IDM-1578: Refactor execution of helpers
Some helpers will be used from other services. Changes name to
clarify that this is execution of helpers and not a generic exec
function.
